LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/worker-import.tentative-expected.txt

CONSOLE MESSAGE: TypeError: WebAssembly modules are not supported in workers yet.

FAIL Testing import of WebAssembly from JavaScript worker Script error.
